Moved frontend VLA RT routines to VLA RT frontend file
